session_log: carve replay.rs (pure move)
The inverse of the typed writer methods: session_log_entry_to_app_event
and its helpers (u32_from_data, read_event_file_span,
read_model_response_content, agent_output_chunks_by_id,
parse_session_attached_message) — the JSONL -> AppEvent replay path used
by dashboard replay and resume rehydration. The rt_* round-trip test
suite rides along; read_events/read_last_event stay in mod.rs as shared
fixtures (pub(crate)-widened).
